local/pyproject.toml
[project]
name = "local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.0",
]
